+++ tests/fkooman/OAuth/Server/TokenIntrospectionServiceTest.php
@@ -25,7 +25,10 @@ use fkooman\Json\Json;
 class TokenIntrospectionServiceTest extends PHPUnit_Framework_TestCase
 {
     /** @var fkooman\OAuth\Server\PdoStorage */
-    protected $storage;
+    private $storage;
+
+    /** @var fkooman\OAuth\Server\Entitlements */
+    private $entitlements;
 
     public function setUp()
     {
@@ -37,46 +40,34 @@ class TokenIntrospectionServiceTest extends PHPUnit_Framework_TestCase
             )
         );
         $this->storage->initDatabase();
-
-        $clientData = new ClientData(
-            array(
-                "id" => "testclient",
-                "name" => "Simple Test Client",
-                "description" => "Client for unit testing",
-                "secret" => null,
-                "icon" => null,
-                "allowed_scope" => "read",
-                "disable_user_consent" => false,
-                "contact_email" => "foo@example.org",
-                "redirect_uri" => "http://localhost/php-oauth/unit/test.html",
-                "type" => "token"
+        $this->storage->addClient(
+            new ClientData(
+                array(
+                    "id" => "testclient",
+                    "name" => "Simple Test Client",
+                    "description" => "Client for unit testing",
+                    "secret" => null,
+                    "icon" => null,
+                    "allowed_scope" => "read",
+                    "disable_user_consent" => false,
+                    "contact_email" => "foo@example.org",
+                    "redirect_uri" => "http://localhost/php-oauth/unit/test.html",
+                    "type" => "token"
+                )
             )
         );
-
-        $this->storage->addClient($clientData);
-
-        $resourceOwnerOne = array(
-            "id" => "fkooman",
-            "entitlement" => array("urn:x-foo:service:access", "urn:x-bar:privilege:admin"),
-            "ext" => array(),
-        );
-        $resourceOwnerTwo = array(
-            "id" => "frko",
-            "entitlement" => array(),
-            "ext" => array(),
-        );
-
-        $this->storage->updateResourceOwner(new MockResourceOwner($resourceOwnerOne));
-        $this->storage->updateResourceOwner(new MockResourceOwner($resourceOwnerTwo));
-
         $this->storage->storeAccessToken("foo", time(), "testclient", "fkooman", "foo bar", 1234);
         $this->storage->storeAccessToken("bar", time(), "testclient", "frko", "a b c", 1234);
+
+        $this->entitlements = new Entitlements(
+            dirname(dirname(dirname(__DIR__))) . '/data/entitlements.json'
+        );
     }
 
     public function testGetTokenIntrospection()
     {
         $h = new Request("https://auth.example.org/introspect?token=foo", "GET");
-        $t = new TokenIntrospectionService($this->storage);
+        $t = new TokenIntrospectionService($this->storage, $this->entitlements);
         $response = $t->run($h);
         $this->assertEquals(200, $response->getStatusCode());
         $this->assertRegexp('|{"active":true,"exp":[0-9]+,"iat":[0-9]+,"scope":"foo bar","client_id":"testclient","sub":"fkooman","token_type":"bearer","x-entitlement":"urn:x-foo:service:access urn:x-bar:privilege:admin"}|', Json::encode($response->getContent()));
@@ -86,7 +77,7 @@ class TokenIntrospectionServiceTest extends PHPUnit_Framework_TestCase
     {
         $h = new Request("https://auth.example.org/introspect", "POST");
         $h->setPostParameters(array("token" => "foo"));
-        $t = new TokenIntrospectionService($this->storage);
+        $t = new TokenIntrospectionService($this->storage, $this->entitlements);
         $response = $t->run($h);
         $this->assertEquals(200, $response->getStatusCode());
         $this->assertRegexp('{"active":true,"exp":[0-9]+,"iat":[0-9]+,"scope":"foo bar","client_id":"testclient","sub":"fkooman","token_type":"bearer","x-entitlement":"urn:x-foo:service:access urn:x-bar:privilege:admin"}', Json::encode($response->getContent()));
@@ -96,7 +87,7 @@ class TokenIntrospectionServiceTest extends PHPUnit_Framework_TestCase
     {
         $h = new Request("https://auth.example.org/introspect", "POST");
         $h->setPostParameters(array("token" => "bar"));
-        $t = new TokenIntrospectionService($this->storage);
+        $t = new TokenIntrospectionService($this->storage, $this->entitlements);
         ;
         $response = $t->run($h);
         $this->assertEquals(200, $response->getStatusCode());
@@ -106,7 +97,7 @@ class TokenIntrospectionServiceTest extends PHPUnit_Framework_TestCase
     public function testMissingGetTokenIntrospection()
     {
         $h = new Request("https://auth.example.org/introspect?token=foobar", "GET");
-        $t = new TokenIntrospectionService($this->storage);
+        $t = new TokenIntrospectionService($this->storage, $this->entitlements);
         ;
         $response = $t->run($h);
         $this->assertEquals(200, $response->getStatusCode());
@@ -120,7 +111,7 @@ class TokenIntrospectionServiceTest extends PHPUnit_Framework_TestCase
     public function testUnsupportedMethod()
     {
         $h = new Request("https://auth.example.org/introspect?token=foobar", "DELETE");
-        $t = new TokenIntrospectionService($this->storage);
+        $t = new TokenIntrospectionService($this->storage, $this->entitlements);
         ;
         $t->run($h);
     }